Add support for specifying proofer logic with a symbol
**WHY**
Most of the time the vendor implementation will want the actual
proofing logic specified in an instance method for simplicity and
ease of testing, resulting in the block provided to `proof` just
calling that method.

**HOW**
Alloe the implementation to provide a symbol that names the method
to call on the implementation, instead of a block, and execute that
method.